% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/kmr_functions.R
\name{anonymize_sids}
\alias{anonymize_sids}
\title{Anonymize Participant IDs}
\usage{
anonymize_sids(df, subject_column_label)
}
\arguments{
\item{df}{Data frame with a column of participant ids that you want to anonymize.}

\item{subject_column_label}{The column name of your participant id variable.}
}
\description{
This function gives participants IDs from 1:length(participants)
}
\examples{
anonymize_sids()
}
\keyword{anonymize}

