% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/apa_manuscript.R
\name{apa_manuscript}
\alias{apa_manuscript}
\title{APA submission format (PDF)}
\usage{
apa_manuscript(keep_tex = TRUE, includes = NULL)
}
\arguments{
\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\description{
Template for creating a handout according to APA
}

