% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/kmr_functions.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Boostrap Confidence Intervals}
\usage{
bootstrap_ci(x, .fun = NULL, bound, level, na.rm = T)
}
\arguments{
\item{x}{A vector of numeric values for which you need a boostrapped confidence interval}

\item{.fun}{The function that you want to be bootsrtrapped}

\item{bound}{Which bound of the confidence interval: "upper" vs. "lower"}

\item{level}{Level of confidence (numeric)}
}
\description{
This function computes 95 percent confidence intervals via non-parametric bootstrap.
}
\examples{
# returns a numeric with either the lower and upper bound of the CI
boostrap_ci(x = df$accuracy, .fun = mean, bound = "upper", level = 0.975, na.rm=T)
}
\keyword{bootstrap}
\keyword{confidence}
\keyword{interval}

