% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/kmr_functions.R
\name{compute_ci}
\alias{compute_ci}
\title{Boostrap Confidence Intervals}
\usage{
compute_ci(x)
}
\arguments{
\item{x}{A vector of numeric values for which you need a boostrapped confidence interval}
}
\description{
This function computes 95 percent confidence intervals via non-parametric bootstrap.
}
\examples{
# returns a named numeric with length 2 that has lower and upper bound of CI
compute_ci(df$accuracy)
}
\keyword{bootstrap}
\keyword{confidence}
\keyword{interval}

