% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/els_article.R
\name{els_manuscript}
\alias{els_manuscript}
\title{ElS submission format (PDF)}
\usage{
els_manuscript(keep_tex = TRUE, includes = NULL)
}
\arguments{
\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\description{
Template for creating a handout according to Elsevier
}

