% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalpred.R
\name{evalpred}
\alias{evalpred}
\title{Evaluate KMR}
\usage{
evalpred(ypred, ytest, type.measure = c("ci", "mse"))
}
\arguments{
\item{ypred}{Matrix or list of matrices (eg. each corresp to a lambda) of predicted reponses of cell lines x tasks.}

\item{ytest}{Matrix of true responses of cell lines x tasks.}

\item{type.measure}{Measure type of evaluation. The default is \code{type.measure="ci"} the concordance index. Other options are \code{type.measure="mse"} the mean squared error.}
}
\description{
Evaluate prediction performance of KMR.
}
