% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausskernel.R
\name{gausskernel}
\alias{gausskernel}
\title{Gaussian RBF kernel}
\usage{
gausskernel(x, sigma = 1, x2)
}
\arguments{
\item{x}{Input matrix of covariates}

\item{sigma}{Bandwidth of the Gaussian RBF kernel (default=1)}

\item{x2}{(optional) A second matrix of covariates}
}
\value{
If \code{x2} is provided, returns the kernel matrix between samples
  in x and samples in x2. If \code{x2} is not provided, returns the kernel
  Gram matrix of x versus itself.
}
\description{
Compute the Gaussian RBF kernel
}
