% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.kmr.R
\name{plot.cv.kmr}
\alias{plot.cv.kmr}
\title{Plot method for a "\code{cv.kmr}" object}
\usage{
\method{plot}{cv.kmr}(x)
}
\arguments{
\item{x}{Fitted "\code{cv.kmr}" model object}
}
\description{
Similar to other \code{plot} methods, this function plots pred perf over regularization parameter for a "\code{cv.kmr}" model object.
}
