% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv.kmr.R
\name{predict.cv.kmr}
\alias{predict.cv.kmr}
\title{Make prediction with a "\code{cv.kmr}" object}
\usage{
\method{predict}{cv.kmr}(object, newx, lambda = "lambda.opt", ...)
}
\arguments{
\item{object}{Fitted "\code{cv.kmr}" model object.}

\item{newx}{Matrix of new values for \code{x} or kernel matrix for new values
crossing old values for \code{x}, at which predictions are to be made.}

\item{lambda}{Value(s) of the regularization parameter, a single scalar or a
sequence of values, or "\code{lambda.opt}" which allows different lambdas 
for each task tuned by cross-validation. Default is "\code{lambda.opt}".}

\item{...}{Not used. Other arguments to predict.}
}
\value{
A matrix of predicted values for the new samples (in rows) and all
  tasks (in columns), corresponding to the regularization parameter
  \code{lambda}. If \code{lambda} is a list of specified values, the function 
  returns a list of matrices, corresponding to the predictions for the different 
  values of \code{lambda} in the list. If \code{lambda} is "\code{lambda.opt}",
  a single matrix returned too.
}
\description{
Similar to other \code{predict} methods, this function predicts fitted values
from a fitted "\code{cv.kmr}" object.
}
\note{
Note that "\code{lambda.opt}" will allow different lambdas for each task tuned
  by cross-validation; otherwise, specified value(s) of lambda will be fixed for all tasks.
}
\seealso{
\code{\link{cv.kmr}}
}
