% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_functions.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{x}{Vector of numeric values}
}
\description{
This function computes the standard error of the mean removing NAs.
}
\examples{
sem()
}
\keyword{sem}

