% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmr_functions.R
\name{to.n}
\alias{to.n}
\title{Convert to numeric}
\usage{
to.n(x)
}
\arguments{
\item{x}{A non numeric value}
}
\description{
This function converts a value to a character than a numeric
}
\keyword{numeric}
\keyword{to}

