% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.kmr.R
\name{cv.kmr}
\alias{cv.kmr}
\title{Cross-validation for KMR}
\usage{
cv.kmr(x, y, kx_type = c("linear", "gaussian", "precomputed"),
  kx_option = list(sigma = 1), kt_type = c("multitask", "empirical",
  "precomputed"), kt_option = list(alpha = 1), lambda = 10^(-5:5),
  type.measure = c("ci", "mse"), nfolds = 3, nrepeats = 1,
  seed = 9182456, mc.cores = 1)
}
\arguments{
\item{x}{\code{x} matrix as in \code{kmr}.}

\item{y}{Reponse matrix \code{y} as in \code{kmr}.}

\item{kx_type}{Kernel type for observations as in \code{kmr}.}

\item{kx_option}{Optional list of parameters for the observation kernel as in \code{kmr}.}

\item{kt_type}{Kernel type for tasks as in \code{kmr}.}

\item{kt_option}{Optional list of parameters for the task kernel as in \code{kmr}.}

\item{lambda}{Sequence of values for lambda that must be tested. Default is 10^(-5:5).}

\item{type.measure}{Loss to use for cross-validation. The default is \code{type.measure="ci"} the concordance index. Other options are \code{type.measure="mse"} the mean squared error.}

\item{nfolds}{Number of folds for cross-validation. Default is 3.}

\item{nrepeats}{Number of times the k-fold cross-validation is performed Default is 1.}

\item{seed}{A seed number for the random number generator (useful to have the same CV splits).}
}
\value{
An object of class \code{"cv.kmr"}, which can then be used to make predictions for the different tasks on new observations as being a list containing the following useful slots:
\item{meanCV}{A matrix of CV performance scores of dim ntask x nlambda.}
\item{bestlambda}{A vector of lambdas of length ntask, each corresp to the underlying min CV score.}
\item{lambda}{Lambda grid to tune over.}
\item{type.measure}{Measure type.}
}
\description{
Does a k-fold cross-validation for \code{kmr}, and returns performance and optimal values for the regularization parameter \code{lambda}.
}
