% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalpred.R
\name{evalpred}
\alias{evalpred}
\title{Evaluate KMR}
\usage{
evalpred(ypred, ytest, type.measure = c("ci", "mse", "cor"))
}
\arguments{
\item{ypred}{Matrix or list of matrices (eg. each corresp to a lambda) of predicted reponses of cell lines x tasks.}

\item{ytest}{Matrix of true responses of cell lines x tasks.}

\item{type.measure}{Character indicating the measure type of evaluation. Possible options are \code{"ci"} (concordance index, default), \code{"mse"} (mean squared error), \code{"cor"} (pearson correlation).}
}
\description{
Evaluate prediction performance of KMR.
}
