% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausskernel.R
\name{gausskernel}
\alias{gausskernel}
\title{Gaussian RBF kernel}
\usage{
gausskernel(x, sigma = 1, x2)
}
\arguments{
\item{x}{Input matrix of covariates with samples in rows.}

\item{sigma}{Bandwidth of the Gaussian RBF kernel. Default is 1.}

\item{x2}{(Optional) a second matrix of covariates with samples in rows.}
}
\value{
If \code{x2} is provided, returns the kernel matrix crossing samples
  in \code{x} and samples in \code{x2}. If \code{x2} is not provided, returns the kernel
  Gram matrix of \code{x} versus itself.
}
\description{
Computes the Gaussian RBF kernel k(x,x') = exp(-|x-x'|^2/(2*sigma^2)).
}
